requirements.txt
Flask~=2.0.2
ipdb
datetime
Flask-SQLAlchemy
PyJWT
Werkzeug~=2.0.2
pymysql
cryptography
Flask-Caching
